+++ packages/crawl_engine/src/crawl_engine/downloads/__init__.py
@@ -1,3 +1,4 @@
+from crawl_engine.downloads.archive_extractor import ArchiveExtractor
 from crawl_engine.downloads.attachment_downloader import AttachmentDownloader
 
-__all__ = ["AttachmentDownloader"]
+__all__ = ["AttachmentDownloader", "ArchiveExtractor"]